+++ ios/Classes/controller/OnPlaylistsController.swift
@@ -10,13 +10,19 @@ class OnPlaylistsController {
         self.result = result
     }
     
+    // Due the [IOS] limitation, for now we can only create playlists.
     func createPlaylist() {
+        //
         let playlistName = args["playlistName"] as! String
         let playlistAuthor = args["playlistAuthor"] as? String
         let playlistDesc = args["playlistDesc"] as? String
+        
+        //
         let playlistMetadata = MPMediaPlaylistCreationMetadata.init(name: playlistName)
         playlistMetadata.authorDisplayName = playlistAuthor
         playlistMetadata.descriptionText = playlistDesc ?? ""
+        
+        //
         MPMediaLibrary().getPlaylist(with: UUID.init(), creationMetadata: playlistMetadata, completionHandler: { playlist, error in
             //A little second to create the playlist and Flutter UI update
             sleep(1)