+++ ios/Classes/query/OnArtistQuery.swift
@@ -6,15 +6,20 @@ class OnArtistsQuery {
     var result: FlutterResult
     
     init(call: FlutterMethodCall, result: @escaping FlutterResult) {
+        // To make life easy, add all arguments inside a map.
         self.args = call.arguments as! [String: Any]
         self.result = result
     }
     
     func queryArtists() {
-//        let sortType = args["sortType"] as! Int
-        
+        // Choose the type(To match android side, let's call "cursor").
         let cursor = MPMediaQuery.artists()
-//        cursor.groupingType = checkArtistSortType(sortType)
+        
+        // We don't need to define a sortType here. [IOS] only support
+        // the [Artist]. The others will be sorted "manually" using
+        // [formatSongList] before send to Dart.
+        
+        // Query everything in background for a better performance.
         loadArtists(cursor: cursor.collections)
     }
     
@@ -22,12 +27,18 @@ class OnArtistsQuery {
         DispatchQueue.global(qos: .userInitiated).async {
             var listOfArtists: [[String: Any?]] = Array()
             
+            // For each item(artist) inside this "cursor", take one and "format"
+            // into a [Map<String, dynamic>], all keys are based on [Android]
+            // platforms so, if you change some key, will have to change the [Android] too.
             for artist in cursor {
                 let artistData = loadArtistItem(artist: artist)
                 listOfArtists.append(artistData)
             }
             
+            // After finish the "query", go back to the "main" thread(You can only call flutter
+            // inside the main thread).
             DispatchQueue.main.async {
+                // Here we'll check the "custom" sort and define a order to the list.
                 let finalList = formatArtistList(args: self.args, allArtists: listOfArtists)
                 self.result(finalList)
             }