+++ ios/Classes/query/OnGenresQuery.swift
@@ -6,12 +6,20 @@ class OnGenresQuery {
     var result: FlutterResult
     
     init(call: FlutterMethodCall, result: @escaping FlutterResult) {
+        // To make life easy, add all arguments inside a map.
         self.args = call.arguments as! [String: Any]
         self.result = result
     }
     
     func queryGenres() {
+        // Choose the type(To match android side, let's call "cursor").
         let cursor = MPMediaQuery.genres()
+        
+        // We don't need to define a sortType here. [IOS] only support
+        // the [Artist]. The others will be sorted "manually" using
+        // [formatSongList] before send to Dart.
+        
+        // Query everything in background for a better performance.
         loadGenres(cursor: cursor.collections)
     }
     
@@ -19,12 +27,18 @@ class OnGenresQuery {
         DispatchQueue.global(qos: .userInitiated).async {
             var listOfGenres: [[String: Any?]] = Array()
             
+            // For each item(genre) inside this "cursor", take one and "format"
+            // into a [Map<String, dynamic>], all keys are based on [Android]
+            // platforms so, if you change some key, will have to change the [Android] too.
             for genre in cursor {
                 let genreData = loadGenreItem(genre: genre)
                 listOfGenres.append(genreData)
             }
             
+            // After finish the "query", go back to the "main" thread(You can only call flutter
+            // inside the main thread).
             DispatchQueue.main.async {
+                // Here we'll check the "custom" sort and define a order to the list.
                 let finalList = formatGenreList(args: self.args, allGenres: listOfGenres)
                 self.result(finalList)
             }