+++ ios/Classes/utils/OnDeviceInfo.swift
@@ -1,5 +1,7 @@
 import Flutter
 
+// TODO: Add more specific return to [Model].
+// This method will get some basic information about platform.
 public func queryDeviceInfo(result: FlutterResult) {
     let device = UIDevice.current
     let deviceData: [String: Any] = [